% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLocationMapping.R
\name{createLocationMapping}
\alias{createLocationMapping}
\title{Create Location Mapping}
\usage{
createLocationMapping(country = "", city = "", lat_long = FALSE)
}
\arguments{
\item{country}{Country name/s for creation scope}

\item{city}{City name/s for creation scope}

\item{lat_long}{True - retrieve coordinates, False - No coordinates}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 11 DF where n is the number of locations in the scope: id, name, slug, city_ID, \cr
city_Name, city_slug, country_ID, country_Name, country_Slug, latitude, longitude
}
\description{
Creates a dataframe listing all locations in a specified scope
}
\examples{
\dontrun{ createLocationMapping("United States", "New York", True)}

}
