% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccountByID.R
\name{getAccountByID}
\alias{getAccountByID}
\title{Get Account By ID}
\usage{
getAccountByID(ID)
}
\arguments{
\item{ID}{An Instagram ID for a media post}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
"1705564303372666424"
}
\description{
Gets the account information for an Instagram ID
}
\examples{
\dontrun{ getCommentsByMediaCode("1579076894773971876", 100)}

}
