% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommentsByMediaCode.R
\name{getCommentsByMediaCode}
\alias{getCommentsByMediaCode}
\title{Get Comments By Media Code}
\usage{
getCommentsByMediaCode(code, n = 10, maxID = "", ...)
}
\arguments{
\item{code}{An Instagram shortcode for a media post}

\item{n}{The number of comments to return}

\item{maxID}{An identifier for a comment that indicates where to start searching}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 6 dataframe - id, text, created_at, owner.id, owner.profile_pic_url, owner.username
}
\description{
Gets the first n comments for a media with a given Instagram shortcode
}
\examples{
\dontrun{ getCommentsByMediaCode("W0IL2cujb3", 100)}

}
