% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommentsByMediaID.R
\name{getCommentsByMediaID}
\alias{getCommentsByMediaID}
\title{Get Comments By Media ID}
\usage{
getCommentsByMediaID(mediaID, n = 10, maxID = "", ...)
}
\arguments{
\item{n}{The number of comments to return}

\item{maxID}{An identifier for a comment that indicates where to start searching}

\item{...}{Additional options passed to a shinyAppDir}

\item{ID}{An Instagram ID for a media post}
}
\value{
n x 6 dataframe - id, text, created_at, owner.id, owner.profile_pic_url, owner.username
}
\description{
Gets the first n comments for a media with a given ID
}
\examples{
\dontrun{ getCommentsByMediaID("1579076894773971876", 100)}

}
