% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentTopMediaByLocationID.R
\name{getCurrentTopMediaByLocationID}
\alias{getCurrentTopMediaByLocationID}
\title{Get Current Top Media By Location ID}
\usage{
getCurrentTopMediaByLocationID(locationID, ...)
}
\arguments{
\item{locationID}{An Instagram ID for a location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
9x14 DF : id, thumbnail_src, thumbnail_resources, is_video, code, date, \cr
display_src, video_views, caption, dimensions.height, dimensions.width, \cr
owner.id, comments.count, likes.count
}
\description{
Gets the top 9 media posts for a location with the given location ID.
}
\examples{
\dontrun{ getCurrentTopMediaByLocationID("212988663")}

}
