% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentTopMediaByLocationMapping.R
\name{getCurrentTopMediaByLocationMapping}
\alias{getCurrentTopMediaByLocationMapping}
\title{Get Current Top Media By Location ID}
\usage{
getCurrentTopMediaByLocationMapping(mapping, ...)
}
\arguments{
\item{mapping}{A location mapping following the format of createLocationMapping()}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
(9*n)x14 DF where n is the number of locations in mapping: \cr
id, thumbnail_src, thumbnail_resources, is_video, code, date, \cr
display_src, video_views, caption, dimensions.height, dimensions.width, \cr
owner.id, comments.count, likes.count
}
\description{
Gets the top 9 media posts for a location with the given location ID.
}
\examples{
\dontrun{
mapping <- createLocationMapping("United States", "New York")
getCurrentTopMediaByLocationMapping(mapping)
}

}
