% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentTopMediaByTag.R
\name{getCurrentTopMediaByTag}
\alias{getCurrentTopMediaByTag}
\title{#'@title Get Current Top Media By Tag}
\usage{
getCurrentTopMediaByTag(tag, ...)
}
\arguments{
\item{tag}{A character string representing the hashtag to be used}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
9x14 DF : id, shortcode, taken_at_timestamp, display_url, thumbnail_src, \cr
thumbnail_resources, is_video, video_view_count, edge_media_to_caption.edges, \cr
edge_media_to_comment.count, dimensions.height, dimensions.wdith, \cr
edge_liked_by.count, owner.id
}
\description{
Gets the top 9 media posts with a given hashtag.
}
\examples{
\dontrun{ getCurrentTopMediaByTag("AirForce")}

}
