% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikesByMediaCode.R
\name{getLikesByMediaCode}
\alias{getLikesByMediaCode}
\title{Get Likes By Media Code}
\usage{
getLikesByMediaCode(code, n = 10, maxID = "", ...)
}
\arguments{
\item{code}{An Instagram shortcode for a media post}

\item{n}{The number of comments to return}

\item{maxID}{An identifier for a comment that indicates where to start searching}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 7 dataframe - id, username,full_name,profile_pic_url, \cr
is_verified_followed_by_viewer,requested_by_viewer
}
\description{
Gets the first n likes for a media with a given Instagram shortcode
}
\examples{
\dontrun{ getLikesByMediaCode("W0IL2cujb3", 100)}

}
