% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikesByMediaID.R
\name{getLikesByMediaID}
\alias{getLikesByMediaID}
\title{Get Likes By Media ID}
\usage{
getLikesByMediaID(ID, n = 10, maxID = "", ...)
}
\arguments{
\item{ID}{An Instagram ID for a media post}

\item{n}{The number of comments to return}

\item{maxID}{An identifier for a comment that indicates where to start searching}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 7 dataframe - id, username,full_name,profile_pic_url, \cr
is_verified_followed_by_viewer,requested_by_viewer
}
\description{
Gets the first n likes for a media with a given Instagram shortcode
}
\examples{
\dontrun{ getLikesByMediaID("1579076894773971876", 100)}

}
