% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationByID.R
\name{getLocationByID}
\alias{getLocationByID}
\title{Get Location by ID}
\usage{
getLocationByID(locationID, ...)
}
\arguments{
\item{locationID}{An Instagram ID for a location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
1 x 6 dataframe - id, name, has_public_page, lat, lng, slug
}
\description{
Returns information about a location with the given ID
}
\examples{
\dontrun{ getLocationByID("212988663")}

}
