% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationsInRange.R
\name{getLocationsInRange}
\alias{getLocationsInRange}
\title{Get Current Top Media By Location ID}
\usage{
getLocationsInRange(mapping, r, lat, long, ...)
}
\arguments{
\item{mapping}{A location mapping in the form generated by createLocationMapping}

\item{r}{Radius of search from give coordinates}

\item{lat}{Double for latitude of search center}

\item{long}{Double for longitude of search center}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 9 DF where n is the number of locations in the scope: \cr
id, name, slug, city_ID, city_Name, city_slug,country_ID, country_Name, country_Slug
}
\description{
Gets the top 9 media posts for a location with the given location ID.
}
\examples{
\dontrun{
mapping <- createLocationMapping("United States", "New York", TRUE)

lat <- 40.7484
long <- 73.9857
radius <- 0.5 #miles

smaller_mapping <- getLocationsInRange(mapping, radius, lat, long)
}

}
