% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMediaByLocationID.R
\name{getMediaByLocationID}
\alias{getMediaByLocationID}
\title{Get Media By Location ID}
\usage{
getMediaByLocationID(locationID, n = 12, maxID = "", ...)
}
\arguments{
\item{locationID}{A Instagram location ID}

\item{n}{The number of media posts to return}

\item{maxID}{Identifier to specify query location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 15 dataframe: comments_disabled, id, thumbnail_src, thumbnail_resources, is_video, \cr
code, date, display_src, video_views, caption, dimension.height, \cr
dimensions.width, owner.id, comments.count, likes.count
}
\description{
Gets the n most recent posts for a location with the given location ID
}
\examples{
\dontrun{ getMediaByLocationID("212988663", 50)}

}
