% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMediaByLocationMapping.R
\name{getMediaByLocationMapping}
\alias{getMediaByLocationMapping}
\title{Get Media By Location Mapping}
\usage{
getMediaByLocationMapping(mapping, n = 12, ...)
}
\arguments{
\item{mapping}{A location mapping following the format of createLocationMapping()}

\item{n}{The number of results to retrieve for each location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
(n*m) x 17 dataframe where m is the number of locations in mapping: \cr
comments_disabled, id, thumbnail_src, is_video, code, date, display_src, \cr
video_views, caption, dimension.height, dimensions.width, \cr
edge_media_preview_like.count, owner.id, comments.count, \cr
likes.count, location_name, location_id
}
\description{
Gets the n most recent media posts for all locations in  a location mapping
}
\examples{
\dontrun
{
mapping <- createLocationMapping("United States", "New York")
getMediaByLocationMapping(mapping)
}

}
