% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMediaByTag.R
\name{getMediaByTag}
\alias{getMediaByTag}
\title{Get Media By Tag}
\usage{
getMediaByTag(tag, n = 20, maxID = "", ...)
}
\arguments{
\item{tag}{A hashtag}

\item{n}{The number of media posts to return}

\item{maxID}{Identifier to specify query location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 15 dataframe: comments_disabled, id, thumbnail_src, thumbnail_resources, is_video, \cr
code, date, display_src, video_views, caption, dimension.height, \cr
dimensions.width, owner.id, comments.count, likes.count
}
\description{
Gets the n most recent posts with the given hashtag
}
\examples{
\dontrun{ getMediaByTag("Batman", 50)}

}
