% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMediaByUsername.R
\name{getMediaByUsername}
\alias{getMediaByUsername}
\title{Get Media By Username}
\usage{
getMediaByUsername(username, n = 12, maxID = "", ...)
}
\arguments{
\item{username}{An Instagram account's username}

\item{n}{The number of media posts to return}

\item{maxID}{Identifier to specify query location}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 17 dataframe of media information: _typename, id, \cr
comments_disabled, geting_info, media_preview, thumbnail_src, \cr
thumbnail_resources, is_video, code, date, display_src, caption, \cr
dimensions.height, dimensions.width, owner.id, comments.count, likes.count
}
\description{
Gets the n most recent posts from a user with the given username
}
\examples{
\dontrun{ getMediaByUsername("officialrickastley", 50)}

}
