% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch App}
\usage{
launchApp(dir, ...)
}
\arguments{
\item{dir}{A file directory for the app}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
N/A
}
\description{
Launches the Shiny App for the instaExtract Package
}
\examples{
\dontrun{ launchApp()}

}
