% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAccountsByUsername.R
\name{searchAccountsByUsername}
\alias{searchAccountsByUsername}
\title{Search Accounts By Username}
\usage{
searchAccountsByUsername(username, ...)
}
\arguments{
\item{username}{A username to search for}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
nx11 dataframe where n is number returned: pk (owner.id), username, full_name, \cr
is_private, profile_pic_url, profile_pic_id, is_verified, has_anonymous_profile_picture, \cr
ollower_count, byline, mutual_followers_count
}
\description{
Gets the results for accounts returned by searching the given username
}
\examples{
\dontrun{ searchAccountsByUsername("AirForce")}

}
