% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTagsbyTag.R
\name{searchTagsByTag}
\alias{searchTagsByTag}
\title{Search Tags by Tag}
\usage{
searchTagsByTag(tag, ...)
}
\arguments{
\item{tag}{A hashtag to search for}

\item{...}{Additional options passed to a shinyAppDir}
}
\value{
n x 3 dataframe where is n is the number of results: colnames : name, id, media_count
}
\description{
Gets the results for hashtags returned by searching the given tag
}
\examples{
\dontrun{ searchTagsByTag("Air")}

}
